third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/worker-src/dedicated-child.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/testharness-helper.js"></script>
<meta http-equiv="Content-Security-Policy" content="child-src http://127.0.0.1:8000 blob:">
<script>
  var url = new URL("/security/contentSecurityPolicy/resources/ping.js", document.baseURI).toString();
  assert_worker_is_loaded(url, "Same-origin dedicated worker allowed by host-source expression.");

  var b = new Blob(["postMessage('ping');"], {type: "text/javascript"});
  var url = URL.createObjectURL(b);
  assert_worker_is_loaded(url, "blob: dedicated worker allowed by 'blob:'.");
</script>
